requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
dash==4.1.0
Flask==3.1.3
idna==3.11
importlib_metadata==9.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.19.0
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
plotly==6.7.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
requests==2.33.1
retrying==1.4.2
setuptools==82.0.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
zipp==3.23.0
